+++ cm_pbio.c
@@ -127,7 +127,7 @@ static int
 CMpbio_send_format_request ARGS((char *format_ID, int format_ID_length,
 				 CMConnection conn, int cond));
 static void CM_pbio_query ARGS((CMConnection conn, CMTransport trans,
-				char *buffer, int length));
+				char *buffer, long length));
 
 static int
 request_in_pending(CManager cm, void *format_ID, int format_id_length)
@@ -533,7 +533,7 @@ byte_swap(char *data, int size)
 }
 
 static void
-CM_pbio_query(CMConnection conn, CMTransport trans, char *buffer, int length)
+CM_pbio_query(CMConnection conn, CMTransport trans, char *buffer, long length)
 {
     struct pbio_exchange_msg tmp_msg;
     struct pbio_exchange_msg *msg;
